Remove @Version annotations and redundant logs
This commit removes the @Version annotations from the Resource and User classes, as well as from the ResourceDTO class. It also eliminates unnecessary logging statements within the UserService and ResourceService classes to improve readability and performance. Enhancements were made to the find methods to ensure they handle null cases and duplicates more effectively.